[feature](merge-cloud) Fix cloud mode can't drop backends and refactoring cloud system info
