+++ fe/fe-core/src/main/java/org/apache/doris/cloud/catalog/CloudClusterChecker.java
@@ -52,8 +52,11 @@ import java.util.stream.Collectors;
 public class CloudClusterChecker extends MasterDaemon {
     private static final Logger LOG = LogManager.getLogger(CloudClusterChecker.class);
 
-    public CloudClusterChecker() {
+    private CloudSystemInfoService cloudSystemInfoService;
+
+    public CloudClusterChecker(CloudSystemInfoService cloudSystemInfoService) {
         super("cloud cluster check", Config.cloud_cluster_check_interval_second * 1000L);
+        this.cloudSystemInfoService = cloudSystemInfoService;
     }
 
     /**
@@ -117,7 +120,7 @@ public class CloudClusterChecker extends MasterDaemon {
                     b.setTagMap(newTagMap);
                     toAdd.add(b);
                 }
-                Env.getCurrentSystemInfo().updateCloudBackends(toAdd, new ArrayList<>());
+                cloudSystemInfoService.updateCloudBackends(toAdd, new ArrayList<>());
             }
         );
     }
@@ -133,15 +136,15 @@ public class CloudClusterChecker extends MasterDaemon {
                 LOG.debug("begin to drop clusterId: {}", delId);
                 List<Backend> toDel =
                         new ArrayList<>(finalClusterIdToBackend.getOrDefault(delId, new ArrayList<>()));
-                Env.getCurrentSystemInfo().updateCloudBackends(new ArrayList<>(), toDel);
+                cloudSystemInfoService.updateCloudBackends(new ArrayList<>(), toDel);
                 // del clusterName
-                String delClusterName = Env.getCurrentSystemInfo().getClusterNameByClusterId(delId);
+                String delClusterName = cloudSystemInfoService.getClusterNameByClusterId(delId);
                 if (delClusterName.isEmpty()) {
                     LOG.warn("can't get delClusterName, clusterId: {}, plz check", delId);
                     return;
                 }
                 // del clusterID
-                Env.getCurrentSystemInfo().dropCluster(delId, delClusterName);
+                cloudSystemInfoService.dropCluster(delId, delClusterName);
             }
         );
     }
@@ -202,12 +205,12 @@ public class CloudClusterChecker extends MasterDaemon {
                 // change all be's cluster_name
                 currentBes.forEach(b -> b.setCloudClusterName(newClusterName));
                 // update clusterNameToId
-                Env.getCurrentSystemInfo().updateClusterNameToId(newClusterName, currentClusterName, cid);
+                cloudSystemInfoService.updateClusterNameToId(newClusterName, currentClusterName, cid);
                 // update tags
                 currentBes.forEach(b -> Env.getCurrentEnv().getEditLog().logModifyBackend(b));
             }
 
-            String currentClusterStatus = Env.getCurrentSystemInfo().getCloudStatusById(cid);
+            String currentClusterStatus = cloudSystemInfoService.getCloudStatusById(cid);
 
             // For old versions that do no have status field set
             ClusterStatus clusterStatus = cp.hasClusterStatus() ? cp.getClusterStatus() : ClusterStatus.NORMAL;
@@ -286,7 +289,7 @@ public class CloudClusterChecker extends MasterDaemon {
                 continue;
             }
 
-            Env.getCurrentSystemInfo().updateCloudBackends(toAdd, toDel);
+            cloudSystemInfoService.updateCloudBackends(toAdd, toDel);
         }
     }
 
@@ -299,7 +302,7 @@ public class CloudClusterChecker extends MasterDaemon {
 
     private void checkFeNodesMapValid() {
         LOG.debug("begin checkFeNodesMapValid");
-        Map<String, List<Backend>> clusterIdToBackend = Env.getCurrentSystemInfo().getCloudClusterIdToBackend();
+        Map<String, List<Backend>> clusterIdToBackend = cloudSystemInfoService.getCloudClusterIdToBackend();
         Set<String> clusterIds = new HashSet<>();
         Set<String> clusterNames = new HashSet<>();
         clusterIdToBackend.forEach((clusterId, bes) -> {
@@ -313,7 +316,7 @@ public class CloudClusterChecker extends MasterDaemon {
             });
         });
 
-        Map<String, String> nameToId = Env.getCurrentSystemInfo().getCloudClusterNameToId();
+        Map<String, String> nameToId = cloudSystemInfoService.getCloudClusterNameToId();
         nameToId.forEach((clusterName, clusterId) -> {
             if (!clusterIdToBackend.containsKey(clusterId)) {
                 LOG.warn("impossible, somewhere err, clusterId {}, clusterName {}, clusterNameToIdMap {}",
@@ -404,7 +407,7 @@ public class CloudClusterChecker extends MasterDaemon {
     }
 
     private void getCloudBackends() {
-        Map<String, List<Backend>> clusterIdToBackend = Env.getCurrentSystemInfo().getCloudClusterIdToBackend();
+        Map<String, List<Backend>> clusterIdToBackend = cloudSystemInfoService.getCloudClusterIdToBackend();
         //rpc to ms, to get mysql user can use cluster_id
         // NOTE: rpc args all empty, use cluster_unique_id to get a instance's all cluster info.
         Cloud.GetClusterResponse response = CloudSystemInfoService.getCloudCluster("", "", "");
@@ -443,14 +446,14 @@ public class CloudClusterChecker extends MasterDaemon {
             LOG.warn("diff cluster has exception, {}", e.getMessage(), e);
 
         }
-        LOG.info("daemon cluster get cluster info succ, current cloudClusterIdToBackendMap: {}",
-                Env.getCurrentSystemInfo().getCloudClusterIdToBackend());
+        LOG.info("daemon cluster get cluster info succ, current cloudClusterIdToBackendMap: {} clusterNameToId {}",
+                cloudSystemInfoService.getCloudClusterIdToBackend(), cloudSystemInfoService.getCloudClusterNameToId());
     }
 
     private void updateCloudMetrics() {
         // Metric
-        Map<String, List<Backend>> clusterIdToBackend = Env.getCurrentSystemInfo().getCloudClusterIdToBackend();
-        Map<String, String> clusterNameToId = Env.getCurrentSystemInfo().getCloudClusterNameToId();
+        Map<String, List<Backend>> clusterIdToBackend = cloudSystemInfoService.getCloudClusterIdToBackend();
+        Map<String, String> clusterNameToId = cloudSystemInfoService.getCloudClusterNameToId();
         for (Map.Entry<String, String> entry : clusterNameToId.entrySet()) {
             long aliveNum = 0L;
             List<Backend> bes = clusterIdToBackend.get(entry.getValue());
