+++ fe/fe-core/src/main/java/org/apache/doris/cloud/system/CloudSystemInfoService.java
@@ -20,9 +20,13 @@ package org.apache.doris.cloud.system;
 import org.apache.doris.catalog.Env;
 import org.apache.doris.catalog.ReplicaAllocation;
 import org.apache.doris.cloud.proto.Cloud;
+import org.apache.doris.cloud.proto.Cloud.ClusterPB;
+import org.apache.doris.cloud.proto.Cloud.InstanceInfoPB;
 import org.apache.doris.cloud.rpc.MetaServiceProxy;
 import org.apache.doris.common.Config;
 import org.apache.doris.common.DdlException;
+import org.apache.doris.common.FeConstants;
+import org.apache.doris.common.Pair;
 import org.apache.doris.ha.FrontendNodeType;
 import org.apache.doris.metric.MetricRepo;
 import org.apache.doris.resource.Tag;
@@ -32,6 +36,7 @@ import org.apache.doris.system.Frontend;
 import org.apache.doris.system.SystemInfoService;
 import org.apache.doris.thrift.TStorageMedium;
 
+import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Maps;
 import org.apache.logging.log4j.LogManager;
@@ -41,12 +46,29 @@ import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Map.Entry;
 import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.locks.ReentrantLock;
 import java.util.stream.Collectors;
 
 public class CloudSystemInfoService extends SystemInfoService {
     private static final Logger LOG = LogManager.getLogger(CloudSystemInfoService.class);
+
+    // TODO(gavin): use {clusterId -> List<BackendId>} instead to reduce risk of inconsistency
+    // use exclusive lock to make sure only one thread can change clusterIdToBackend and clusterNameToId
+    protected ReentrantLock lock = new ReentrantLock();
+
+    // for show cluster and cache user owned cluster
+    // mysqlUserName -> List of ClusterPB
+    private Map<String, List<ClusterPB>> mysqlUserNameToClusterPB = ImmutableMap.of();
+    // clusterId -> List<Backend>
+    protected Map<String, List<Backend>> clusterIdToBackend = new ConcurrentHashMap<>();
+    // clusterName -> clusterId
+    protected Map<String, String> clusterNameToId = new ConcurrentHashMap<>();
+
+    private InstanceInfoPB.Status instanceStatus;
 
     @Override
     public Map<Tag, List<Long>> selectBackendIdsForReplicaCreation(
@@ -61,7 +83,7 @@ public class CloudSystemInfoService extends SystemInfoService {
      * Gets cloud cluster from remote with either clusterId or clusterName
      *
      * @param clusterName cluster name
-     * @param clusterId cluster id
+     * @param clusterId   cluster id
      * @return
      */
     public static Cloud.GetClusterResponse getCloudCluster(String clusterName, String clusterId, String userName) {
@@ -86,7 +108,7 @@ public class CloudSystemInfoService extends SystemInfoService {
             LOG.info("nothing to do");
             return;
         }
-        Set<String> existedBes = idToBackendRef.entrySet().stream().map(i -> i.getValue())
+        Set<String> existedBes = idToBackendRef.values().stream()
                 .map(i -> i.getHost() + ":" + i.getHeartbeatPort())
                 .collect(Collectors.toSet());
         LOG.debug("deduplication existedBes={}", existedBes);
@@ -229,7 +251,7 @@ public class CloudSystemInfoService extends SystemInfoService {
 
 
     public static synchronized void updateFrontends(List<Frontend> toAdd,
-                                             List<Frontend> toDel) throws DdlException {
+                                                    List<Frontend> toDel) throws DdlException {
         LOG.debug("updateCloudFrontends toAdd={} toDel={}", toAdd, toDel);
         String masterIp = Env.getCurrentEnv().getMasterHost();
         for (Frontend fe : toAdd) {
@@ -248,5 +270,167 @@ public class CloudSystemInfoService extends SystemInfoService {
             LOG.info("dropped cloud frontend={} ", fe);
         }
     }
-}
 
+    public void replayDropBackend(Backend backend) {
+        super.replayDropBackend(backend);
+        List<Backend> toDel = new ArrayList<>();
+        toDel.add(backend);
+        updateCloudClusterMap(new ArrayList<>(), toDel);
+    }
+
+    public boolean availableBackendsExists() {
+        if (FeConstants.runningUnitTest) {
+            return true;
+        }
+        if (null == clusterNameToId || clusterNameToId.isEmpty()) {
+            return false;
+        }
+        return clusterIdToBackend != null && !clusterIdToBackend.isEmpty()
+            && clusterIdToBackend.values().stream().anyMatch(list -> list != null && !list.isEmpty());
+    }
+
+    public boolean containClusterName(String clusterName) {
+        return clusterNameToId.containsKey(clusterName);
+    }
+
+    public List<Backend> getBackendsByClusterName(final String clusterName) {
+        String clusterId = clusterNameToId.getOrDefault(clusterName, "");
+        if (clusterId.isEmpty()) {
+            return new ArrayList<>();
+        }
+        return clusterIdToBackend.get(clusterId);
+    }
+
+    public List<Backend> getBackendsByClusterId(final String clusterId) {
+        return clusterIdToBackend.getOrDefault(clusterId, new ArrayList<>());
+    }
+
+    public List<String> getCloudClusterIds() {
+        return new ArrayList<>(clusterIdToBackend.keySet());
+    }
+
+    public String getCloudStatusByName(final String clusterName) {
+        String clusterId = clusterNameToId.getOrDefault(clusterName, "");
+        if (Strings.isNullOrEmpty(clusterId)) {
+            // for rename cluster or dropped cluster
+            LOG.warn("cant find clusterId by clusterName {}", clusterName);
+            return "";
+        }
+        return getCloudStatusById(clusterId);
+    }
+
+    public String getCloudStatusById(final String clusterId) {
+        return clusterIdToBackend.getOrDefault(clusterId, new ArrayList<>())
+            .stream().map(Backend::getCloudClusterStatus).findFirst().orElse("");
+    }
+
+    public void updateClusterNameToId(final String newName,
+                                      final String originalName, final String clusterId) {
+        lock.lock();
+        clusterNameToId.remove(originalName);
+        clusterNameToId.put(newName, clusterId);
+        lock.unlock();
+    }
+
+    public String getClusterNameByClusterId(final String clusterId) {
+        String clusterName = "";
+        for (Map.Entry<String, String> entry : clusterNameToId.entrySet()) {
+            if (entry.getValue().equals(clusterId)) {
+                clusterName = entry.getKey();
+                break;
+            }
+        }
+        return clusterName;
+    }
+
+    public void dropCluster(final String clusterId, final String clusterName) {
+        lock.lock();
+        clusterNameToId.remove(clusterName, clusterId);
+        clusterIdToBackend.remove(clusterId);
+        lock.unlock();
+    }
+
+    public List<String> getCloudClusterNames() {
+        return new ArrayList<>(clusterNameToId.keySet());
+    }
+
+    // Return the ref of concurrentMap clusterIdToBackend
+    // It should be thread-safe to iterate.
+    // reference: https://stackoverflow.com/questions/3768554/is-iterating-concurrenthashmap-values-thread-safe
+    public Map<String, List<Backend>> getCloudClusterIdToBackend() {
+        return clusterIdToBackend;
+    }
+
+    public String getCloudClusterIdByName(String clusterName) {
+        return clusterNameToId.get(clusterName);
+    }
+
+    public ImmutableMap<Long, Backend> getCloudIdToBackend(String clusterName) {
+        String clusterId = clusterNameToId.get(clusterName);
+        if (Strings.isNullOrEmpty(clusterId)) {
+            LOG.warn("cant find clusterId, this cluster may be has been dropped, clusterName={}", clusterName);
+            return ImmutableMap.of();
+        }
+        List<Backend> backends = clusterIdToBackend.get(clusterId);
+        Map<Long, Backend> idToBackend = Maps.newHashMap();
+        for (Backend be : backends) {
+            idToBackend.put(be.getId(), be);
+        }
+        return ImmutableMap.copyOf(idToBackend);
+    }
+
+    // Return the ref of concurrentMap clusterNameToId
+    // It should be thread-safe to iterate.
+    // reference: https://stackoverflow.com/questions/3768554/is-iterating-concurrenthashmap-values-thread-safe
+    public Map<String, String> getCloudClusterNameToId() {
+        return clusterNameToId;
+    }
+
+    public Map<String, List<ClusterPB>> getMysqlUserNameToClusterPb() {
+        return mysqlUserNameToClusterPB;
+    }
+
+    public void updateMysqlUserNameToClusterPb(Map<String, List<ClusterPB>> m) {
+        mysqlUserNameToClusterPB = m;
+    }
+
+    public List<Pair<String, Integer>> getCurrentObFrontends() {
+        List<Frontend> frontends = Env.getCurrentEnv().getFrontends(FrontendNodeType.OBSERVER);
+        List<Pair<String, Integer>> frontendsPair = new ArrayList<>();
+        for (Frontend frontend : frontends) {
+            frontendsPair.add(Pair.of(frontend.getHost(), frontend.getEditLogPort()));
+        }
+        return frontendsPair;
+    }
+
+    public Cloud.GetInstanceResponse getCloudInstance() {
+        Cloud.GetInstanceRequest.Builder builder = Cloud.GetInstanceRequest.newBuilder();
+        builder.setCloudUniqueId(Config.cloud_unique_id);
+        final Cloud.GetInstanceRequest pRequest = builder.build();
+        Cloud.GetInstanceResponse response;
+        try {
+            response = MetaServiceProxy.getInstance().getInstance(pRequest);
+            return response;
+        } catch (RpcException e) {
+            LOG.warn("rpcToGetInstance exception: {}", e.getMessage());
+        }
+        return null;
+    }
+
+    public InstanceInfoPB.Status getInstanceStatus() {
+        return this.instanceStatus;
+    }
+
+    public void setInstanceStatus(InstanceInfoPB.Status instanceStatus) {
+        LOG.debug("fe set instance status {}", instanceStatus);
+        if (this.instanceStatus != instanceStatus) {
+            LOG.info("fe change instance status from {} to {}", this.instanceStatus, instanceStatus);
+        }
+        this.instanceStatus = instanceStatus;
+    }
+
+    public static void waitForAutoStart(final String clusterName) throws DdlException {
+        // TODO: merge from cloud.
+        throw new DdlException("Env.waitForAutoStart unimplemented");
+    }
+}
