+++ fe/fe-core/src/main/java/org/apache/doris/system/SystemInfoService.java
@@ -23,10 +23,6 @@ import org.apache.doris.catalog.Database;
 import org.apache.doris.catalog.DiskInfo;
 import org.apache.doris.catalog.Env;
 import org.apache.doris.catalog.ReplicaAllocation;
-import org.apache.doris.cloud.proto.Cloud;
-import org.apache.doris.cloud.proto.Cloud.ClusterPB;
-import org.apache.doris.cloud.proto.Cloud.InstanceInfoPB;
-import org.apache.doris.cloud.rpc.MetaServiceProxy;
 import org.apache.doris.common.AnalysisException;
 import org.apache.doris.common.Config;
 import org.apache.doris.common.DdlException;
@@ -36,11 +32,9 @@ import org.apache.doris.common.Status;
 import org.apache.doris.common.UserException;
 import org.apache.doris.common.io.CountingDataOutputStream;
 import org.apache.doris.common.util.NetUtils;
-import org.apache.doris.ha.FrontendNodeType;
 import org.apache.doris.metric.MetricRepo;
 import org.apache.doris.qe.ConnectContext;
 import org.apache.doris.resource.Tag;
-import org.apache.doris.rpc.RpcException;
 import org.apache.doris.thrift.TNodeInfo;
 import org.apache.doris.thrift.TPaloNodesInfo;
 import org.apache.doris.thrift.TStatusCode;
@@ -69,9 +63,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicLong;
-import java.util.concurrent.locks.ReentrantLock;
 import java.util.stream.Collectors;
 
 public class SystemInfoService {
@@ -88,21 +80,8 @@ public class SystemInfoService {
 
     protected volatile ImmutableMap<Long, Backend> idToBackendRef = ImmutableMap.of();
     protected volatile ImmutableMap<Long, AtomicLong> idToReportVersionRef = ImmutableMap.of();
-    // TODO(gavin): use {clusterId -> List<BackendId>} instead to reduce risk of inconsistency
-    // use exclusive lock to make sure only one thread can change clusterIdToBackend and clusterNameToId
-    protected ReentrantLock lock = new ReentrantLock();
-
-    // for show cluster and cache user owned cluster
-    // mysqlUserName -> List of ClusterPB
-    private Map<String, List<ClusterPB>> mysqlUserNameToClusterPB = ImmutableMap.of();
-    // clusterId -> List<Backend>
-    protected Map<String, List<Backend>> clusterIdToBackend = new ConcurrentHashMap<>();
-    // clusterName -> clusterId
-    protected Map<String, String> clusterNameToId = new ConcurrentHashMap<>();
 
     private volatile ImmutableMap<Long, DiskInfo> pathHashToDiskInfoRef = ImmutableMap.of();
-
-    private InstanceInfoPB.Status instanceStatus;
 
     public static class HostInfo implements Comparable<HostInfo> {
         public String host;
@@ -183,114 +162,6 @@ public class SystemInfoService {
         }
     };
 
-    public boolean availableBackendsExists() {
-        if (FeConstants.runningUnitTest) {
-            return true;
-        }
-        if (null == clusterNameToId || clusterNameToId.isEmpty()) {
-            return false;
-        }
-        return clusterIdToBackend != null && !clusterIdToBackend.isEmpty()
-               && clusterIdToBackend.values().stream().anyMatch(list -> list != null && !list.isEmpty());
-    }
-
-    public boolean containClusterName(String clusterName) {
-        return clusterNameToId.containsKey(clusterName);
-    }
-
-    public List<Backend> getBackendsByClusterName(final String clusterName) {
-        String clusterId = clusterNameToId.getOrDefault(clusterName, "");
-        if (clusterId.isEmpty()) {
-            return new ArrayList<>();
-        }
-        return clusterIdToBackend.get(clusterId);
-    }
-
-    public List<Backend> getBackendsByClusterId(final String clusterId) {
-        return clusterIdToBackend.getOrDefault(clusterId, new ArrayList<>());
-    }
-
-    public List<String> getCloudClusterIds() {
-        return new ArrayList<>(clusterIdToBackend.keySet());
-    }
-
-    public String getCloudStatusByName(final String clusterName) {
-        String clusterId = clusterNameToId.getOrDefault(clusterName, "");
-        if (Strings.isNullOrEmpty(clusterId)) {
-            // for rename cluster or dropped cluster
-            LOG.warn("cant find clusterId by clusterName {}", clusterName);
-            return "";
-        }
-        return getCloudStatusById(clusterId);
-    }
-
-    public String getCloudStatusById(final String clusterId) {
-        return clusterIdToBackend.getOrDefault(clusterId, new ArrayList<>())
-            .stream().map(Backend::getCloudClusterStatus).findFirst().orElse("");
-    }
-
-    public void updateClusterNameToId(final String newName,
-            final String originalName, final String clusterId) {
-        lock.lock();
-        clusterNameToId.remove(originalName);
-        clusterNameToId.put(newName, clusterId);
-        lock.unlock();
-    }
-
-    public String getClusterNameByClusterId(final String clusterId) {
-        String clusterName = "";
-        for (Map.Entry<String, String> entry : clusterNameToId.entrySet()) {
-            if (entry.getValue().equals(clusterId)) {
-                clusterName = entry.getKey();
-                break;
-            }
-        }
-        return clusterName;
-    }
-
-    public void dropCluster(final String clusterId, final String clusterName) {
-        lock.lock();
-        clusterNameToId.remove(clusterName, clusterId);
-        clusterIdToBackend.remove(clusterId);
-        lock.unlock();
-    }
-
-    public List<String> getCloudClusterNames() {
-        return new ArrayList<>(clusterNameToId.keySet());
-    }
-
-    // Return the ref of concurrentMap clusterIdToBackend
-    // It should be thread-safe to iterate.
-    // reference: https://stackoverflow.com/questions/3768554/is-iterating-concurrenthashmap-values-thread-safe
-    public Map<String, List<Backend>> getCloudClusterIdToBackend() {
-        return clusterIdToBackend;
-    }
-
-    public String getCloudClusterIdByName(String clusterName) {
-        return clusterNameToId.get(clusterName);
-    }
-
-    public ImmutableMap<Long, Backend> getCloudIdToBackend(String clusterName) {
-        String clusterId = clusterNameToId.get(clusterName);
-        if (Strings.isNullOrEmpty(clusterId)) {
-            LOG.warn("cant find clusterId, this cluster may be has been dropped, clusterName={}", clusterName);
-            return ImmutableMap.of();
-        }
-        List<Backend> backends = clusterIdToBackend.get(clusterId);
-        Map<Long, Backend> idToBackend = Maps.newHashMap();
-        for (Backend be : backends) {
-            idToBackend.put(be.getId(), be);
-        }
-        return ImmutableMap.copyOf(idToBackend);
-    }
-
-    // Return the ref of concurrentMap clusterNameToId
-    // It should be thread-safe to iterate.
-    // reference: https://stackoverflow.com/questions/3768554/is-iterating-concurrenthashmap-values-thread-safe
-    public Map<String, String> getCloudClusterNameToId() {
-        return clusterNameToId;
-    }
-
     public static TPaloNodesInfo createAliveNodesInfo() {
         TPaloNodesInfo nodesInfo = new TPaloNodesInfo();
         SystemInfoService systemInfoService = Env.getCurrentSystemInfo();
@@ -299,23 +170,6 @@ public class SystemInfoService {
             nodesInfo.addToNodes(new TNodeInfo(backend.getId(), 0, backend.getHost(), backend.getBrpcPort()));
         }
         return nodesInfo;
-    }
-
-    public Map<String, List<ClusterPB>> getMysqlUserNameToClusterPb() {
-        return mysqlUserNameToClusterPB;
-    }
-
-    public void updateMysqlUserNameToClusterPb(Map<String, List<ClusterPB>> m) {
-        mysqlUserNameToClusterPB = m;
-    }
-
-    public List<Pair<String, Integer>> getCurrentObFrontends() {
-        List<Frontend> frontends = Env.getCurrentEnv().getFrontends(FrontendNodeType.OBSERVER);
-        List<Pair<String, Integer>> frontendsPair = new ArrayList<>();
-        for (Frontend frontend : frontends) {
-            frontendsPair.add(Pair.of(frontend.getHost(), frontend.getEditLogPort()));
-        }
-        return frontendsPair;
     }
 
     // for deploy manager
@@ -1127,36 +981,5 @@ public class SystemInfoService {
 
     public long aliveBECount() {
         return idToBackendRef.values().stream().filter(Backend::isAlive).count();
-    }
-
-    public Cloud.GetInstanceResponse getCloudInstance() {
-        Cloud.GetInstanceRequest.Builder builder =
-                Cloud.GetInstanceRequest.newBuilder();
-        builder.setCloudUniqueId(Config.cloud_unique_id);
-        final Cloud.GetInstanceRequest pRequest = builder.build();
-        Cloud.GetInstanceResponse response;
-        try {
-            response = MetaServiceProxy.getInstance().getInstance(pRequest);
-            return response;
-        } catch (RpcException e) {
-            LOG.warn("rpcToGetInstance exception: {}", e.getMessage());
-        }
-        return null;
-    }
-
-    public InstanceInfoPB.Status getInstanceStatus() {
-        return this.instanceStatus;
-    }
-
-    public void setInstanceStatus(InstanceInfoPB.Status instanceStatus) {
-        LOG.debug("fe set instance status {}", instanceStatus);
-        if (this.instanceStatus != instanceStatus) {
-            LOG.info("fe change instance status from {} to {}", this.instanceStatus, instanceStatus);
-        }
-        this.instanceStatus = instanceStatus;
-    }
-
-    public synchronized void updateCloudBackends(List<Backend> toAdd, List<Backend> toDel) {
-        LOG.warn("Not cloud mode, should not be here");
     }
 }
